creat bus trip, bus schedule

+++ src/main/java/com/pbl6/VehicleBookingRental/user/service/impl/UtilitySeviceImpl.java
@@ -64,6 +64,9 @@ public class UtilitySeviceImpl implements UtilityService {
     @Override
     public void deleteUtility(int idUtility) throws IdInvalidException {
         Utility utilityDb = this.getUtilityById(idUtility);
+        if(utilityDb.getBuses()!= null){
+            throw new RuntimeException("Cannot delete this utility, it has buses");
+        }
         this.s3Service.deleteFile(utilityDb.getImage());
 
         this.utilityRepository.delete(utilityDb);
